+++ src/main/java/com/github/javafaker/service/RandomService.java
@@ -17,4 +17,8 @@ public class RandomService {
     public int nextInt(int n) {
         return RandomUtils.nextInt(random, n);
     }
+
+    public double nextDouble() {
+        return RandomUtils.nextDouble(random);
+    }
 }